Refactor SRA tests into their own file (#10775)
* Begin refactoring SRA tests out and to use realWin

* Fix test

* Fixed lint

* responded to feedback

* Responded to feedback

* Fix minor mistake

* Responded to feedback

* Switch to using createAndAppendAdElement

* Fix lint
